requirements.txt
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.5.0
cuda-toolkit==12.9.1
cycler==0.12.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.17.1.4
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile==1.15.1.6
nvidia-cufile-cu12==1.14.1.1
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.28.9
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.4.5
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.9.79
packaging==26.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
setuptools==81.0.0
six==1.17.0
sympy==1.14.0
torch==2.11.0+cu129
torchaudio==2.11.0+cu129
torchvision==0.26.0+cu129
triton==3.6.0
typing_extensions==4.15.0
